Update README to match the one on learn-co-students
The README on the batch learn-co-curriculum seems to have a slightly more-updated README, that does not include the section about starting the server. 
Also, the following line has been corrected: `Run rails db:migrate to create a schema, models, and controllers` to point to what `rake db:migrate` actually does.